there might be better way to do this

+++ src/hardware/hardwareOutputDevice.h
@@ -13,7 +13,7 @@ public:
     virtual bool configure(std::unordered_map<string, string> settings) = 0;
 
     //Set a hardware channel output. Value is 0.0 to 1.0 for no to max output.
-    virtual void setChannelData(int channel, float value) = 0;
+    virtual void setChannelData(int channel, int value) = 0;
     
     //Return the number of output channels supported by this device.
     virtual int getChannelCount() = 0;
