+++ src/hardware/virtualOutputDevice.h
@@ -37,7 +37,7 @@ public:
     virtual bool configure(std::unordered_map<string, string> settings);
 
     //Set a hardware channel output. Value is 0.0 to 1.0 for no to max output.
-    virtual void setChannelData(int channel, float value);
+    virtual void setChannelData(int channel, int value);
     
     //Return the number of output channels supported by this device.
     virtual int getChannelCount();
